+++ packages/bitbucket/mobile-header/src/components/MobileHeader.js
@@ -5,7 +5,7 @@ import MenuIcon from '@atlaskit/icon/glyph/menu';
 import * as styles from '../styled';
 
 type MobileHeaderProps = {
-  navigation: (isOpen: boolean) => Node,
+  navigation?: (isOpen: boolean) => Node,
   sidebar: (isOpen: boolean) => Node,
   onNavigationOpen: () => void,
   onDrawerClose: () => void,
@@ -13,6 +13,7 @@ type MobileHeaderProps = {
   secondaryContent?: Node,
   pageHeading: Node,
   menuIconLabel: string,
+  customMenu?: Node,
 };
 
 type MobileHeaderState = {
@@ -47,7 +48,7 @@ class MobileHeader extends PureComponent<MobileHeaderProps, MobileHeaderState> {
   renderSlider = (
     isOpen: boolean,
     isAnimating: boolean,
-    renderFn: (isOpen: boolean) => Node,
+    renderFn?: (isOpen: boolean) => Node,
     onTransitionEnd: Function,
     side: string = 'left',
   ) => (
@@ -56,25 +57,29 @@ class MobileHeader extends PureComponent<MobileHeaderProps, MobileHeaderState> {
       onTransitionEnd={onTransitionEnd}
       side={side}
     >
-      {(isOpen || isAnimating) && renderFn(isOpen)}
+      {(isOpen || isAnimating) && renderFn && renderFn(isOpen)}
     </styles.MobileNavSlider>
   );
 
   render() {
     const { isAnimatingNavigation, isAnimatingSidebar } = this.state;
-    const { drawerState, menuIconLabel } = this.props;
+    const { drawerState, menuIconLabel, customMenu } = this.props;
     const isNavigationOpen = drawerState === 'navigation';
     const isSidebarOpen = drawerState === 'sidebar';
+
+    const menu = customMenu || (
+      <Button
+        appearance="subtle"
+        iconBefore={<MenuIcon label={menuIconLabel} size="large" />}
+        onClick={this.props.onNavigationOpen}
+      />
+    );
 
     return (
       <Fragment>
         <styles.MobilePageHeader>
           <styles.MobilePageHeaderContent>
-            <Button
-              appearance="subtle"
-              iconBefore={<MenuIcon label={menuIconLabel} size="large" />}
-              onClick={this.props.onNavigationOpen}
-            />
+            {menu}
             <styles.PageHeading>{this.props.pageHeading}</styles.PageHeading>
             {this.props.secondaryContent}
           </styles.MobilePageHeaderContent>