+++ packages/editor/renderer/src/__tests__/unit/email/index.ts
@@ -8,7 +8,10 @@ import * as em from '../../__fixtures__/em.adf.json';
 import * as codeBlock from '../../__fixtures__/code-block.adf.json';
 import * as inlineCodeProps from '../../__fixtures__/inline-code-props.adf.json';
 import * as inlineTextProps from '../../__fixtures__/inline-text-props.adf.json';
+import * as panels from '../../__fixtures__/panels.adf.json';
 import * as link from '../../__fixtures__/link.adf.json';
+import * as blockCards from '../../__fixtures__/block-cards.adf.json';
+import * as inlineCards from '../../__fixtures__/inline-cards.adf.json';
 import * as status from '../../__fixtures__/status.adf.json';
 
 const render = (doc: any) => {
@@ -21,8 +24,23 @@ const render = (doc: any) => {
 };
 
 describe('Renderer - EmailSerializer', () => {
+  it('should render block cards correctly', () => {
+    const output = render(blockCards);
+    expect(output).toMatchSnapshot();
+  });
+
+  it('should render inline cards correctly', () => {
+    const output = render(inlineCards);
+    expect(output).toMatchSnapshot();
+  });
+
   it('should render text with em inside of a paragraph correctly', () => {
     const output = render(em);
+    expect(output).toMatchSnapshot();
+  });
+
+  it('should render panels correctly', () => {
+    const output = render(panels);
     expect(output).toMatchSnapshot();
   });