+++ packages/elements/user-picker/src/__tests__/unit/ssr/server-side-hydrate.tsx
@@ -19,6 +19,7 @@ describe('server side rendering and hydration', async () => {
     ['03-single-compact.tsx'],
     ['04-single-subtle.tsx'],
     ['05-single-subtle-and-compact.tsx'],
+    ['09-watchers.tsx'],
     ['11-creatable.tsx'],
   ])('ssr("%s")', async (fileName: string) => {
     await ssr_hydrate(__dirname, `${ExamplesPath}/${fileName}`);
@@ -28,10 +29,9 @@ describe('server side rendering and hydration', async () => {
   });
 
   test.skip.each([
-    ['06-multi-with-default-values.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3688
-    ['07-multi-with-fixed-values.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3688
-    ['08-single-disabled.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3688
-    ['09-watchers.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3688
+    ['06-multi-with-default-values.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3704
+    ['07-multi-with-fixed-values.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3704
+    ['08-single-disabled.tsx'], // TODO: https://product-fabric.atlassian.net/browse/FS-3704
     ['10-in-a-table-cell.tsx'], // depends on document (DOM)
   ])('ssr("%s")', async (fileName: string) => {
     await ssr_hydrate(__dirname, `${ExamplesPath}/${fileName}`);