+++ packages/elements/task-decision/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/task-decision
 
+## 13.2.0
+- [minor] [b0210d7ccc](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b0210d7ccc):
+
+  - reset jest modules before hydration
+
 ## 13.1.1
 - [patch] [1bcaa1b991](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1bcaa1b991):