+++ packages/navigation/atlassian-switcher/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/atlassian-switcher
 
+## 0.3.1
+- [patch] [57f774683f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/57f774683f):
+
+  - Move @atlaskit/logo to peer dependencies
+
+## 0.3.0
+- [minor] [68443e3d6f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/68443e3d6f):
+
+  - Opsgenie app switching support
+
 ## 0.2.3
 - [patch] [a041506c4d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a041506c4d):